+++ CHANGES.md
@@ -4,6 +4,7 @@ JAMES Examples Module Changes
 Version 0.2.1 (under development)
 ---------------------------------
 
+ - Provided examples for using the analysis tools from the extensions module.
  - Fixed conflicting logback settings in tests. Settings are now inherited from the core module.
 
 Version 0.2 (12/11/2014)